requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pypi_0
aiohttp=3.8.3=pypi_0
aiosignal=1.2.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.1.0=pypi_0
autobahn=22.7.1=pypi_0
automat=20.2.0=pypi_0
blas=1.0=mkl
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py38h0a891b7_1004
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.9.24=ha878542_0
cachetools=5.2.0=pypi_0
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
constantly=15.1.0=pypi_0
cryptography=38.0.1=pypi_0
cudatoolkit=11.6.0=hecad31d_10
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pypi_0
deprecated=1.2.13=pypi_0
expat=2.4.9=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
freetype-py=2.3.0=pypi_0
frozenlist=1.3.1=pypi_0
fsspec=2022.8.2=pypi_0
future=0.18.2=pypi_0
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.12.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.49.1=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
hydra-core=1.2.0=pypi_0
hyperlink=21.0.0=pypi_0
icu=58.2=he6710b0_3
idna=3.4=pyhd8ed1ab_0
imageio=2.22.1=pypi_0
importlib-metadata=5.0.0=pypi_0
importlib-resources=5.9.0=pypi_0
incremental=21.3.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h7f8727e_0
kiwisolver=1.4.2=py38h295c915_0
krb5=1.19.2=hac12032_0
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.4.0=hecacb30_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
liegroups=1.1.0=pypi_0
lxml=4.9.1=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.2=py38h06a4308_0
matplotlib-base=3.5.2=py38hf590b9c_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=6.0.2=pypi_0
munkres=1.1.4=py_0
ncurses=6.3=h5eee18b_3
nettle=3.6=he412f7d_0
networkx=2.2=pypi_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numpy=1.22.4=pypi_0
oauthlib=3.2.1=pypi_0
omegaconf=2.2.3=pypi_0
openh264=2.1.1=h780b84a_0
openssl=1.1.1o=h166bdaf_0
packaging=21.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pillow=9.2.0=py38hace64e9_1
pip=22.2.2=py38h06a4308_0
ply=3.11=py38_0
protobuf=3.19.6=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pyb-utils=0.2.0=pypi_0
pybullet=3.2.5=pypi_0
pycollada=0.6=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydeprecate=0.3.2=pypi_0
pyglet=1.5.27=pypi_0
pyopengl=3.1.0=pypi_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=py38h06a4308_0
pyqt=5.15.7=py38h6a678d5_1
pyqt5=5.15.7=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.11.0=py38h6a678d5_1
pyrender=0.1.45=pypi_0
pysdf=0.1.8=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.12.1=py3.8_cuda11.6_cudnn8.3.2_0
pytorch-lightning=1.7.7=pypi_0
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.1.2=h7f8727e_1
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=pypi_0
setuptools=63.4.1=py38h06a4308_0
sip=6.6.2=py38h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h5082296_0
tacto=0.0.3=pypi_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tifffile=2022.8.12=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.12.1=py38_cu116
torchmetrics=0.10.0=pypi_0
torchvision=0.13.1=py38_cu116
tornado=6.2=py38h5eee18b_0
tqdm=4.64.1=pypi_0
trimesh=3.15.3=pypi_0
twisted=22.8.0=pypi_0
txaio=22.2.1=pypi_0
typing_extensions=4.4.0=pyha770c72_0
urdfpy=0.0.22=pypi_0
urllib3=1.26.11=pyhd8ed1ab_0
vedo=2022.3.0=pypi_0
vtk=9.0.3=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.14.1=pypi_0
wslink=1.8.2=pypi_0
xz=5.2.6=h5eee18b_0
yarl=1.8.1=pypi_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h5eee18b_3
zope-interface=5.4.0=pypi_0
zstd=1.5.2=ha4553b6_0
